Move the abstracted plugin runner code into a separate file
This way main.py contains just the ipa-healthcheck main code
and not the rest of the abstracted code.

Also replace sys.exit(1) with return 1.
